+++ src/main/java/io/github/nocomment1105/mobspecificenchants/enchantments/IEnchantmentAndHelper.java
@@ -1,12 +1,11 @@
 package io.github.nocomment1105.mobspecificenchants.enchantments;
 
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.ItemStack;
 
 public interface IEnchantmentAndHelper
 {
-    default float getEntityAttackDamage(Entity entity, int level)
+    default float getEntityAttackDamage(Entity entity, float level)
     {
         return 0.0f;
     }
